+++ src/tq.go
@@ -205,6 +205,7 @@ func tq_append(channel int, prio int, pp *packet_t) {
 			nettnc_send_packet(channel, pp)
 		}
 
+		ackmode_discard(pp) // rerouted off-radio, not transmitted - drop any pending ACKMODE ack
 		ax25_delete(pp)
 
 		return
@@ -222,6 +223,7 @@ func tq_append(channel int, prio int, pp *packet_t) {
 		dw_printf("original KISS protocol specification.  The solution might be to use\n")
 		dw_printf("a command like \"kissparms -c 1 -p radio\" to set CRC none mode.\n")
 		dw_printf("\n")
+		ackmode_discard(pp) // invalid channel, not transmitted - drop any pending ACKMODE ack
 		ax25_delete(pp)
 
 		return
@@ -254,6 +256,7 @@ func tq_append(channel int, prio int, pp *packet_t) {
 		text_color_set(DW_COLOR_ERROR)
 		dw_printf("Transmit packet queue for channel %d is too long.  Discarding packet.\n", channel)
 		dw_printf("Perhaps the channel is so busy there is no opportunity to send.\n")
+		ackmode_discard(pp) // queue overflow, not transmitted - drop any pending ACKMODE ack
 		ax25_delete(pp)
 
 		return